cleanup documenation component without state
